Added a dropdown (popup) menu component used in the NavDropdown component for a button which opens to a list of links or buttons in a popup menu below the nav (good for a 'User' menu).
Removed some unused styling for a label in the Progress component, as the label is no longer part of the component (it was for a while during development until I realized it's easier to style if you just give it an ID and style it how you need it).

Added a canned styling set for a nav within the footer: 'nav.simple-footer-links' which just shows a few links in a row nicely spaced, and in a column beneath 650px (around four links).