Made the '.modal-body' element something the consumer needs to add if they want. It just provides built-in padding currently in addition to the styles it had in the component.
Added some table styling, including a faint background color on 'th' cells, no borders, and some padding. The more it's used the more it'll probably need to be changed but this is a start.

Added a 'flex-wrap: wrap' utility class.

Incrementing the minor version now, so this is 0.1.0. From now on it'll follow semver a bit better.