Use a bit less of a dull/dark purple color theme for background colors. Swapping it to be a real dark blue sticks out a little less as far as dark themes go.
Removed the purple top border on the footer, it was way too bright and felt like a laser beam in a bad way.

Changed the panel border color from black to a light (transparent) gray. This makes it feel more like a highlight and gives it a hint of more color and contrasts nicely with the blues in buttons.