Lots of changes here and there needed for Overseer. It's a bigger app than any of the other things using sheodox-ui so it needed a lot of things!
I've added a new MenuButton component, it lets you have a popup menu of options that show when you click a button. It's a pattern used elsewhere but now it's an official component because it's tricky getting it to show/hide at the right times. The positioning of something generic like this is a pain with just CSS so I added popper.js to handle positioning the popup.

Tons of other tiny changes:
* Several flexbox related utilities (align-self: start/center/end, and flex: 1 through flex: 5) and font weight/text align utility classes were added.
* Tables can now add a .hoverable class to show a row highlight on hover.
* TH cells have a more fitting background color.
* Labels in input groups are now vertically centered and padded by default.
* Header nav links can now have a .active class which shows a pink border on it to indicate the current page.
* Panels can now get some padding by adding a .panel-body class to it (not just to an element it contains) and the panel border can be applied on all sides by adding .bordered.
* Modals now have a .modal-footer class you can use which will right align a row of buttons.
* The sub panel bg now has a color that more matches the color scheme elsewhere.
* Modals now use the panel border, they were (and probably still are) hard to see on low contrast displays.

Toasts now support progress bars and links. If an href is provided in the toast options it'll make the message into a link. When creating a progress toast you can call the new updateToast method to overwrite any needed options.

This is still a work in progress as I finish rewriting overseer. This will probably cause a few breaking changes around popups!