A couple of bigger changes here. I added a Portal component so things that look like they're not a child of their parent component (popups, modals) actually will be mounted directly to the body. This lets things like the nav dropdown menu and modals not inherit styles from wherever they're triggered, preventing awkward event passing to get a modal to show from an action clicked in the header to make it mount in a safer place.
Also made it so nav dropdowns are positioned with popper instead of some awkward css positioning that happened to get it more or less in the right spot.

Navs dropdowns also used "aria-pressed" when their respective menus are opened instead of a pressed class, it's probably not the most applicable use of aria, but it's closer.