Added a TextInput component, it's a input[type=text] with a floating label that shows more emphasized over the field until the field has a value or is focused.
Also renamed all of the css variables to be prefixed with `shdx-` so they're less likely to conflict with a site's own variables.

Added some styling for disabling checkboxes. It's not currently tested but should work, possibly expect fixes if it doesn't look right.

Added date and time inputs to the input styling, otherwise they looked pretty vanilla.