Typescript rewrite!
This rewrites everything with typescript, and packages it with Sveltekit because it seemed easier than figuring out how to package typings myself. Because of how Sveltekit expects files to be structured the folder structure changed a bit (all exported files need to be in src/lib).

This also adds code formatting with Prettier, and linting with svelte-check, which are both run pre-commit with husky.