Changed the Icon component's padding options
There was a `noPadding` prop, but this would only allow for icons that are prepended, or icon only. If you need to append an icon (like for a "next" button where you want the icon to follow the text for symmetry with a matching "previous" button) you'd need to manually configure padding. There are now three variants (the previous 'variant' prop which changed which icon set in use has been changed to `iconVariant`) for if the icon is prepended/appended/icon-only (equivalent to the old `noPadding={true}`).

This also wasn't using the spacing system defined padding so I changed that.