Misc style changes
Added a .secondary button which includes a background all the time. Some tweaks for consistent disabled button styles etc.

The gray styles all use a css variable for the hue, which can be overridden allowing a cooler/warmer tinge to the background colors.

Fleshed out the demo page a bit more with more demos and widgets