Rewrite progress component
Previously a vanilla progress element was used, with some moz/webkit overrides to style it properly, but that was unwieldy so I've rewritten it with aria stuff so it's just a couple divs with the proper attributes on them.

They have been rethemed too. Previously there was a 1px border and the background on the bar would match the gradient of the border, but the purple of the gradient was kind of dark, and with the progress bar being so short (it has been doubled to 2*spacing height instead of 1*spacing) it was kind of hard to look at. It is now just a bright pink with a semi-transparent gray so it will hopefully still contrast nicely with whatever background it has.

Also npm audit fix