Added extra fieldset variations
The existing 'variant' is now 'size', and 'variant' now changes the style/them that the tabs use between three different styles. `title` is now `legend` as it's more indicative of the html it's creating.

The normal style previously was a tabbed look, but if you have nested fieldsets or the background color is nested fieldsets would be effectively invisible. The box and bordered fieldsets use transparent colors. I would like transparent colors on the tab variant too, but with the css trickery in use to create that tab effect transparent colors would get extra bright wherever there was overlap.

Also made it so the legends don't wrap.