Fixed bordered fieldset legend width
It was generally way too small with the ellipsis styling in there, making almost all legends have an ellipsis, when there is a ton of space. The bordered fieldset doesn't have the same issues with wrapping text as the other variants do, so it can wrap just fine.

Also added a border radius on the box variant.