Fix legend text overflow for the box variant
When you put a max width on a fieldset it was growing out of the container to fit the legend, making the fieldset too large in the process. This just replaces the styling used for the text ellipsis for box variants with something that seems to play better (it is vendor prefixed, but the vendor prefixed version seems to be standardized and well supported).

The tab style still has this problem for now.