New conventional hamburger menu sidebar, demo site sveltekit update
This adds a responsive sidebar openable in two different modes with a button in the header.

This also updates the version of sveltekit in use for packaging and dev. The dev site now breaks down different components into several different pages, with routes and links in the new sidebar component.