Allow sidebar width to be changed, new component containing base sheodox-ui styles
This lets you pass a css width (like '20rem') to the Sidebar component, also the default width is lowered to 20rem because it looked better.

I also made a component `SheodoxUIStyles` that just imports the main stylesheet. For apps that for whatever reason have trouble importing the styles directly this is a workaround.