Sticky position header and sidebar, smaller header logo
This lets pages that scroll still have the sidebar and header visible, without the sidebar awkwardly ending as soon as you scroll a bit.

The header logo also shrinks, as it needs to be smaller for the header sizing to match up correctly with the height of the header as used by the sidebar.