Hover colors on toggle, muted color that matches theme
Adding a lighter color on toggles when hovering so it gives some feedback and some extra distinction between the other options.

Updated the `.muted` class to use a gray color from the theme instead of a random hex, so it adjusts with the gray hue of the page.

This also fixes the misnamed "Radios" page, now that it's more than that.