Allow specifying modal size
This lets you specify a size for modals (height/width as strings). This is useful for modals that contain tabs which might have different sizes, allowing you to specify an ideal size so the modal doesn't move around or resize between tabs, and it will only shrink below that size if there's not enough room on the page.