Allow fieldset padding to be configured, default padding changes with size
The fieldset padding was the same for every size, but it should really scale with the size of the heading. You can also use a css variable `--sx-arg-fieldset-padding` to change what is used.

Also fixed the top padding calculation for box fieldsets, since by default the fieldset title sits on top of the top border it should only need to be pushed down by 50% of the text size, this makes all edges have more consistent padding.