Add alerts, inverted badges, hr color
This adds a new Alert component generally intended for top of the page alert banners.

Added a new variant to badges that  uses a dark color on light background, for when you want more punchy colors.

I updated the color on `hr` elements so it is a mild gray instead of blinding bright white.