Validate links are real before counting it as a link
The parsing function needs to be able to recognize actual links, because otherwise it'll throw errors when rendering. This tries make a new URL object out of it, and if it doesn't parse it just considers it text. Also anytime text segments are added it'll check if the previous segment was also text and concat it, so parsing invalid URLs doesn't create extra text segments for no reason.