Allow customizable header positions
If you have a page that has content positioned with translateY you might have content below the page's size, allowing scrolling but at some point you'd scroll to a point where the sticky header disappears. You need to handle positioning for a position: fixed header so doing this is opt-in.