New LinkTabList component, remove padding variants on Icons
There's a new LinkTabList component, meant to be nav bar where each tab is really just a different page url, nice when the tabs are good UX, but everything's actually on different pages.

I also removed the existing 'variant' on Icon which was just meant to control padding before/after the icon, and renamed `iconVariant` to `variant`. This is because the padding use case is better served by just putting an actual space between the icon and text, and it just causes tons of `variant="icon-only"` props to litter the app.