Better toasts
This adds `success` and `warning` toast variants.

I also made it so toasts only have a message. In practice the title+message concept was weird and usually caused me to duplicate text, or just pass an empty string for the message. Toasts normally don't have both a title and a message, so now sheodox-ui's toasts only have a message.

I also made the info toast use a really dark background color, because it was too close to normal page colors it made it harder to notice most of the time.

Blockquotes have less intense styles now.