Fix outline on focused dropdown menu buttons
To maintain the rounded corners on the menu it used `overflow: hidden;` which worked fine, but had the unintended side effect of making the outlines not show up properly (at max one or two edges would be visible). The border radius is now applied to the top/bottom items in the dropdown so the effect is maintained without causing issues with the focus styling.